soql_agent/requirements.txt
langchain
langgraph
langchain-openai
streamlit
python-dotenv
requests
pydantic
